write file only if not exists

+++ apt_venv/utils.py
@@ -20,21 +20,14 @@ def get_template(filename):
         result = templates.APT_CONF
     return result
 
-def change_dir(dir):
-    debug(2, "moving to directory %s" % dir)
-    try:
-        os.chdir(dir)
-    except OSError as oserror:
-        raise OSError("OSError [%d]: %s at %s" % \
-          (oserror.errno, oserror.strerror, oserror.filename))
-
 def create_file(filename, content):
-    debug(2, "creating file %s" % filename)
-    content = u'%s' % content
-    if content[-1] != '\n':
-        content += '\n'
-    with codecs.open(filename, 'w', 'utf-8') as writer:
-        writer.write(content)
+    if not os.path.isfile(filename):
+        debug(2, "creating file %s" % filename)
+        content = u'%s' % content
+        if content[-1] != '\n':
+            content += '\n'
+        with codecs.open(filename, 'w', 'utf-8') as writer:
+            writer.write(content)
 
 def create_dir(dir):
     if not os.path.isdir(dir):
